+++ app/Apis/Capito/Simplifications.php
@@ -63,10 +63,11 @@ class Simplifications extends Simplification_Base implements Api_Simplifications
 	 * @param string $source_language The source language of the text.
 	 * @param string $target_language The target language of the text.
 	 * @param bool   $is_html Marker if the text contains HTML-Code.
+	 * @param bool   $is_test Marker if this is a rest request.
 	 * @return array<string,int|string> The result as array.
 	 * @noinspection PhpUnused
 	 */
-	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html ): array {
+	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html, bool $is_test = false ): array {
 		// map the languages with its shorthand (e.g. de_DE => de).
 		$source_language = $this->get_api()->get_supported_source_languages()[ $source_language ]['api_value'];
 		$target_language = $this->get_api()->get_supported_target_languages()[ $target_language ]['api_value'];