+++ app/Apis/No_Api/Simplifications.php
@@ -62,9 +62,10 @@ class Simplifications extends Simplification_Base implements Api_Simplifications
 	 * @param string $source_language The source language of the text.
 	 * @param string $target_language The target language of the text.
 	 * @param bool   $is_html Marker if the text contains HTML-Code.
+	 * @param bool   $is_test Marker if this is a rest request.
 	 * @return array<string,int|string> The result as array.
 	 */
-	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html ): array {
+	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html, bool $is_test = false ): array {
 		return array();
 	}
 }