+++ app/Apis/Summ_Ai/Simplifications.php
@@ -55,26 +55,18 @@ class Simplifications extends Simplification_Base implements Api_Simplifications
 	 * @param string $source_language The source language of the text.
 	 * @param string $target_language The target language of the text.
 	 * @param bool   $is_html Marker if the text contains HTML-Code.
+	 * @param bool   $is_test Marker if this is a rest request.
 	 * @return array<string,int|string> The result as array.
 	 */
-	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html ): array {
-		// get the separator setting for target language.
-		$separators = (array) get_option( 'easy_language_summ_ai_target_languages_separator' );
-		$separator  = $separators[ $target_language ];
+	public function call_api( string $text_to_translate, string $source_language, string $target_language, bool $is_html, bool $is_test = false ): array {
+		// get the separator setting for the target language.
+		$enable_separator = (string) get_option( 'easy_language_summ_ai_target_languages_' . $target_language . '_separator' );
 
 		// get the new line setting for target language.
-		$new_lines_array = (array) get_option( 'easy_language_summ_ai_target_languages_new_lines' );
-		$new_lines       = 0;
-		if ( ! empty( $new_lines_array[ $target_language ] ) ) {
-			$new_lines = 1;
-		}
+		$enable_new_lines = absint( get_option( 'easy_language_summ_ai_target_languages_' . $target_language . '_new_line' ) );
 
 		// get the embolden setting for target language.
-		$embolden_negatives_array = (array) get_option( 'easy_language_summ_ai_target_languages_embolden_negative' );
-		$embolden_negative        = 0;
-		if ( ! empty( $embolden_negatives_array[ $target_language ] ) ) {
-			$embolden_negative = 1;
-		}
+		$enable_embolden_negative = absint( get_option( 'easy_language_summ_ai_target_languages_' . $target_language . '_embolden_negative' ) );
 
 		// build request.
 		$request_obj = $this->get_api()->get_request_object();
@@ -82,11 +74,11 @@ class Simplifications extends Simplification_Base implements Api_Simplifications
 		$request_obj->set_token( $this->get_api()->get_token() );
 		$request_obj->set_text( $text_to_translate );
 		$request_obj->set_text_type( $is_html && 1 === absint( get_option( 'easy_language_summ_ai_html_mode' ) ) ? 'html' : 'plain_text' );
-		$request_obj->set_separator( $separator );
-		$request_obj->set_new_lines( $new_lines );
-		$request_obj->set_embolden_negative( $embolden_negative );
+		$request_obj->set_separator( $enable_separator );
+		$request_obj->set_new_lines( $enable_new_lines );
+		$request_obj->set_embolden_negative( $enable_embolden_negative );
 		$request_obj->set_method( 'POST' );
-		$request_obj->set_is_test( $this->get_api()->is_test_mode_active() );
+		$request_obj->set_is_test( $is_test ? true : $this->get_api()->is_test_mode_active() );
 		$request_obj->set_source_language( $source_language );
 		$request_obj->set_target_language( $target_language );
 		/**